131 balance cargo reward system (#143)
* refactor: remove Currency

flattens CurrencyManager

* refactor: change weight from double to int

* feat: added CurrencyRangedManager

for currency with ranges

* refactor: utilise new CurrencyRangedManager

* refactor: change double to int for currency value

* test: added test for remaining methods

for CurrencyRangedManager

* fix: rebase merge conflicts

* style: semantic name switch

* Revert "style: semantic name switch"

This reverts commit a9a4bec7d3d5dfbdb6a70a0a884856dc8b133e9e.

* style: semantic name switch

+++ RailwayCo/Assets/Scripts/Unity/Main/LogicManager.cs
@@ -241,10 +241,10 @@ public class LogicManager : MonoBehaviour
         int exp = _gameManager.GameLogic.User.ExperiencePoint;
         
         CurrencyManager currMgr = _gameManager.GameLogic.User.GetCurrencyManager();
-        double coinVal = currMgr.GetCurrency(CurrencyType.Coin);
-        double noteVal = currMgr.GetCurrency(CurrencyType.Note);
-        double normalCrateVal = currMgr.GetCurrency(CurrencyType.NormalCrate);
-        double specialCrateVal = currMgr.GetCurrency(CurrencyType.SpecialCrate);
+        int coinVal = currMgr.GetCurrency(CurrencyType.Coin);
+        int noteVal = currMgr.GetCurrency(CurrencyType.Note);
+        int normalCrateVal = currMgr.GetCurrency(CurrencyType.NormalCrate);
+        int specialCrateVal = currMgr.GetCurrency(CurrencyType.SpecialCrate);
 
         statsPanel.Find("EXPText").GetComponent<Text>().text = exp.ToString();
         statsPanel.Find("CoinText").GetComponent<Text>().text = coinVal.ToString();
